DatabaseConnect/1-Prep/requirements.txt
# Updated requirements.txt for Python 3.13
asn1crypto>=1.5.1,<2.0.0
certifi>=2025.1.31
cffi>=1.17.1
charset-normalizer>=3.4.1
cryptography>=41.0.7,<42.0.0
filelock>=3.18.0
idna>=3.10
packaging>=24.2
platformdirs>=3.11.0
pycparser>=2.22
PyJWT>=2.10.1
pyOpenSSL>=23.3.0
python-dotenv>=1.0.1
pytz>=2025.1
requests>=2.32.3
snowflake-connector-python>=3.5.0
sortedcontainers>=2.4.0
tomlkit>=0.13.2
typing_extensions>=4.12.2
urllib3>=2.0.0  # Updated from 1.26.20 for better Python 3.13 compatibility